+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.1 (unreleased)
 ----------------
 
+- Export as an actual csv file.
+  [davidjb]
 - Updated the text on the add keywords page.
   [davidjb]
 - First real version. See README for a feature list.